+++ rememB/partyroomapp/views.py
@@ -1,11 +1,10 @@
+from turtle import left
 from userapp.models import User
 from letterapp.models import Letter
 from rest_framework.views import APIView
 from rest_framework.response import Response
 from letterapp.serializers import *
-from userapp.authenticate import SafeJWTAuthentication
-from rest_framework import permissions
-from rest_framework import status
+from userapp.models import User
 
 class UserLetterView(APIView):
     #authentication_classes=[SafeJWTAuthentication]
@@ -15,7 +14,13 @@ class UserLetterView(APIView):
     def get(self,request,userpk):
         #token_user=str(SafeJWTAuthentication.authenticate(self, request)[0])
         #request_user=str(User.objects.all())
-
+        user=User.objects.get(id=userpk)
+        leftDay=User.getDayBefore(str(user.birth))
         user_letters=Letter.objects.filter(user=userpk)
         serializer=LetterSumSerializer(user_letters, many=True)
-        return Response(serializer.data)
+        data={
+            'left_birth':leftDay,
+            'letters':serializer.data
+            
+        }
+        return Response(data)
